:tractor: relocated hashstore to core folder
