+++ moban/core/hashstore.py
