+++ tests/test_hash_store.py
@@ -3,7 +3,7 @@ import sys
 
 from nose import SkipTest
 from moban.externals import file_system
-from moban.hashstore import HashStore
+from moban.core.hashstore import HashStore
 
 
 class TestHashStore:
